+++ ChangeLog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Import setting to create non-legacy animation clips (thanks [@hybridherbst][hybridherbst] for #196) 
+- Support for individual texture transform per texture type (URP, HDRP and Built-in)
 ### Fixed
 - Editor import: Separate textures are only referenced in AssetDatabase (not re-added)
 - Warnings due to conflicting script file names `Animation.cs` and `Camera.cs` (#198)